move chance level function
