+++ recovery/plot_group_lvl.py
@@ -8,31 +8,9 @@ import re
 import sys
 sys.path.append(str(Path(__file__).parents[1]))
 from plot_fns import plot_recoveries, plot_descriptive_adequacy
-from utils import logit, inv_logit
+from utils import logit, inv_logit, chance_level
 
 
-def chance_level(n, alpha = 0.001, p = 0.5):
-    """
-    Calculates the chance level for a given number of trials and alpha level
-
-    Parameters
-    ----------
-    n : int
-        The number of trials.
-    alpha : float
-        The alpha level.
-    p : float
-        The probability of a correct response.
-
-    Returns
-    -------
-    chance_level : float
-        The chance level.
-    """
-    k = binom.ppf(1-alpha, n, p)
-    chance_level = k/n
-    
-    return chance_level
 
 
 def load_simulated(path : Path) -> dict:
